Add configuration parameter for fs devices in autotest
Add Host.fsdevs attribute which contains a dict of name, path pairs for
directories to be mounted into the guest as fs devices. The fsdevs
parameter in the host section of autotest.cfg should contain those pairs
in the format name1:path1,name2:path2 ... Nothing is also possible in
case you don't wanna use fsdevs at all.

Note that the name is also used as mount tag, and should therefore
correspond to those set in the guest-config.nix.