Add support for TCP TSO
This is implemented via a new virtual `send_tso` driver function.
Currently, only the dpdk driver implements it. But because the interface
is designed with a fallback mechanism in mind, we can provide a default
implementation for other drivers which simply fails silently.

The e810 simulator will try to use this for TSO, falling back on its
software TSO implementation if `send_tso` fails.